refactor: extract TranscriptReconciler behind an explicit ReconcileHost
Phase 4 of the engine.ts decomposition (#867). The ~1,950-line transcript
reconciliation cluster (entry-id and content-identity tail anchoring,
append-only read optimization, replay/flood analysis, bootstrap-state
refresh) moves to src/transcript-reconciler.ts. Like rotation, it runs
against a narrow host interface that inventories its engine coupling
(stores, session queue, ingestSingle, bootstrap-import recording) and
takes the SessionRolloverDetector as a direct collaborator. The
checkpoint helpers (BootstrapCheckpointFileState,
checkpointIsPastTranscriptEof) move with it and are re-exported for the
engine's bootstrap path.

Method bodies unchanged; unifying the overlapping bootstrap vs after-turn
reconcile paths is intentionally NOT done here (it is a semantic change
in the hottest bug area; tracked on #867). Tests that spied on the
private methods now spy on engine.transcriptReconciler.

Full suite: 1279 tests passing.
