+++ src/engine.ts
@@ -1,5 +1,4 @@
 import { createHash } from "node:crypto";
-import { readFileSync } from "node:fs";
 import { open, stat } from "node:fs/promises";
 import { join } from "node:path";
 import type { DatabaseSync } from "node:sqlite";
@@ -39,15 +38,15 @@ import { logStartupBannerOnce } from "./startup-banner-log.js";
 import { CompactionTelemetryStore } from "./store/compaction-telemetry-store.js";
 import { CompactionMaintenanceStore } from "./store/compaction-maintenance-store.js";
 import { ConversationStore, type ConversationRecord } from "./store/conversation-store.js";
-import { buildMessageIdentityHash } from "./store/message-identity.js";
 import { FocusBriefStore, type FocusBriefRecord } from "./store/focus-brief-store.js";
 import { SummaryStore, type ContextItemRecord } from "./store/summary-store.js";
 import { createLcmSummarizeFromLegacyParams, FALLBACK_SUMMARY_MARKER, LcmProviderAuthError, LcmSummarySpendLimitError, type LcmSummarizeFn } from "./summarize.js";
 import type { LcmDependencies } from "./types.js";
 import { estimateTokens } from "./estimate-tokens.js";
 import { buildDeterministicFallbackSummary } from "./summary-fallback.js";
-import { getTranscriptEntryId, getTranscriptEntryMeta, readAppendedLeafPathMessages, readLastJsonlEntryBeforeOffset, readLeafPathMessages, readSessionParentSessionReference, readTranscriptHeader } from "./transcript.js";
-import { resolveEpochRoute, selectEntryIdTail, transcriptImportCap, type TranscriptReconcileResult } from "./reconcile-plan.js";
+import { getTranscriptEntryId, readAppendedLeafPathMessages, readLastJsonlEntryBeforeOffset, readLeafPathMessages, readSessionParentSessionReference } from "./transcript.js";
+import { type TranscriptReconcileResult } from "./reconcile-plan.js";
+import { checkpointIsPastTranscriptEof, TranscriptReconciler } from "./transcript-reconciler.js";
 import { AMBIGUOUS_SESSION_KEY_RUNTIME_ROLLOVER_REASON, SessionRolloverDetector } from "./session-rollover.js";
 import {
   SessionRotationService,
@@ -60,14 +59,14 @@ import {
 import { describeAssembledPrefixChange, formatOverflowDiagnosticsForLog, shouldLogOverflowDiagnostics, type AssemblePrefixSnapshot, type BootstrapImportObservation } from "./assemble-debug.js";
 import { appendForkBoundedLiveSuffixWithinBudget, buildDegradedLiveAssembleResult, buildForkBoundedLiveFallback, clampMessagesToSerializedBudget, resolveDeferredAssemblyPressure } from "./assemble-fallback.js";
 import { resolveBootstrapMaxTokens, trimBootstrapMessagesToBudget } from "./bootstrap-budget.js";
-import { batchLooksLikeHeartbeatAckTurn, filterSyntheticHeartbeatMessages, pruneHeartbeatOkTurns } from "./heartbeat-filter.js";
+import { batchLooksLikeHeartbeatAckTurn, pruneHeartbeatOkTurns } from "./heartbeat-filter.js";
 import { appendUncoveredVolatileLiveInputsWithinBudget, isVolatileLiveInputMessage, messageContentCoveredBySummary, resolveProtectedFreshTailAssembledIndexes } from "./live-coverage.js";
-import { buildMessageParts, extractMessageContent, filterPersistableMessages, hasPersistableMessageRole, isLikelyInjectedDeliveryOnlyTranscript, isLikelyInjectedMetadataPreambleRecord, isOpenClawRuntimeContextLeak, toStoredMessage, type StoredMessage } from "./message-content.js";
-import { createBootstrapEntryHash, createLosslessMessageSignature, isBootstrapReplayCandidateMessage, messageIdentity, readBootstrapMessageFromJsonLine } from "./message-signatures.js";
+import { buildMessageParts, extractMessageContent, filterPersistableMessages, hasPersistableMessageRole, isOpenClawRuntimeContextLeak, toStoredMessage } from "./message-content.js";
+import { createBootstrapEntryHash, readBootstrapMessageFromJsonLine } from "./message-signatures.js";
 import { PROMPT_RECALL_MAX_MESSAGES, PROMPT_RECALL_SEARCH_CANDIDATE_LIMIT, buildPromptRecallProjectionFingerprint, extractPromptRecallIdentifiers, extractPromptRecallSnippet, findPromptRecallIdentifierIndex, isPromptRecallEligibleRole, normalizePromptRecallCoverageText, normalizePromptRecallText, renderPromptRecallMessage } from "./prompt-recall.js";
-import { externalizedReplayMetadataMatches, extractPlainToolReplayTextsById, extractRawBlockIdsFromPartMetadata, extractRawBlockSignatureFromPartMetadata, extractRawIdsFromPartMetadata, listTranscriptToolResultEntryIdsByCallId } from "./replay-metadata.js";
+import { listTranscriptToolResultEntryIdsByCallId } from "./replay-metadata.js";
 import { estimateSessionTokenCountForAfterTurn, extractRuntimePromptTokenCount } from "./token-accounting.js";
-import { asRecord, formatDurationMs, isMissingFileError, resolvePositiveInteger, safeString } from "./value-utils.js";
+import { asRecord, formatDurationMs, resolvePositiveInteger } from "./value-utils.js";
 
 type AgentMessage = Parameters<ContextEngine["ingest"]>[0]["message"];
 const LOSSLESS_AGENT_RUN_REQUIRED_HOST_CAPABILITIES: ContextEngineHostCapability[] = [
@@ -83,19 +82,8 @@ const LOSSLESS_SUBAGENT_SPAWN_REQUIRED_HOST_CAPABILITIES: ContextEngineHostCapab
 ];
 const MAX_PREVIOUS_ASSEMBLED_SNAPSHOTS = 100;
 const FORK_BOUNDED_BOOTSTRAP_REASON = "fork-bounded bootstrap import";
-/**
- * How deep into the conversation tail a flush-lagged runtime row can sit.
- * Flush lag is a same-turn phenomenon (the runtime persisted a row moments
- * before the transcript caught up), so matches deeper than this are legacy
- * unstamped rows, not flush lag.
- */
-const FLUSH_LAG_ADOPTION_TAIL_WINDOW = 16;
 const CONTEXT_ENGINE_PROJECTION_EPOCH_VERSION = "summary-prefix-v1";
 const DEFERRED_ASSEMBLY_DEGRADED_PRESSURE_RATIO = 0.75;
-type BootstrapCheckpointFileState = {
-  lastProcessedOffset: number;
-  lastSeenSize: number;
-};
 type CompactionExecutionParams = {
   conversationId: number;
   sessionId: string;
@@ -176,15 +164,6 @@ function buildFocusProjectionKey(brief?: FocusBriefRecord | null): string | null
   return hash.digest("hex").slice(0, 32);
 }
 
-function checkpointIsPastTranscriptEof(
-  checkpoint: BootstrapCheckpointFileState | null | undefined,
-  fileSize: number,
-): boolean {
-  if (!checkpoint) {
-    return false;
-  }
-  return checkpoint.lastProcessedOffset > fileSize || checkpoint.lastSeenSize > fileSize;
-}
 
 const TRANSCRIPT_GC_BATCH_SIZE = 12;
 
@@ -269,22 +248,9 @@ export class LcmContextEngine implements ContextEngine {
   // ── Managed session-file rotation ────────────────────────────────────────
   private readonly sessionRotation: SessionRotationService;
 
-  /** Last file state successfully covered by `reconcileTranscriptTailForAfterTurn`
-   *  slow-path full re-reads, keyed by `${sessionQueueKey}\u0000${sessionFile}`
-   *  (same NUL-escape separator pattern as `messageIdentity`). Long-running
-   *  sessions where the bootstrap checkpoint is missing or path-mismatched
-   *  would otherwise pay O(file-size) on every afterTurn; repeated attempts
-   *  for the same unchanged file state are skipped.
-   *
-   *  Bounded with FIFO eviction at `AFTER_TURN_RECONCILE_KEY_CAP` entries
-   *  so hosts churning through many sessions/files don't accumulate this
-   *  map indefinitely. When the cap is exceeded we drop the oldest entry
-   *  (Map iteration order is insertion order in JS); a session whose
-   *  entry eventually evicts may pay the slow path once again, which is
-   *  acceptable since the bound is well above realistic concurrent-session
-   *  counts. */
-  private afterTurnReconcileFullReadStates = new Map<string, { size: number; mtimeMs: number }>();
-  private static readonly AFTER_TURN_RECONCILE_KEY_CAP = 4096;
+  // ── Transcript/DB tail reconciliation ────────────────────────────────────
+  private readonly transcriptReconciler: TranscriptReconciler;
+
 
   constructor(deps: LcmDependencies, database: DatabaseSync) {
     this.deps = deps;
@@ -444,6 +410,25 @@ export class LcmContextEngine implements ContextEngine {
       compactionConfig,
       this.deps.log,
     );
+    this.transcriptReconciler = new TranscriptReconciler(
+      {
+        config: this.config,
+        db: this.db,
+        deps: this.deps,
+        conversationStore: this.conversationStore,
+        summaryStore: this.summaryStore,
+        shouldIgnoreSession: (params) => this.shouldIgnoreSession(params),
+        resolveSessionQueueKey: (sessionId, sessionKey) =>
+          this.resolveSessionQueueKey(sessionId, sessionKey),
+        withSessionQueue: (queueKey, operation, options) =>
+          this.withSessionQueue(queueKey, operation, options),
+        formatSessionLogContext: (params) => this.formatSessionLogContext(params),
+        recordRecentBootstrapImport: (conversationId, importedMessages, reason) =>
+          this.recordRecentBootstrapImport(conversationId, importedMessages, reason),
+        ingestSingle: (params) => this.ingestSingle(params),
+      },
+      this.sessionRolloverDetector,
+    );
     this.sessionRotation = new SessionRotationService({
       config: this.config,
       db: this.db,
@@ -464,11 +449,11 @@ export class LcmContextEngine implements ContextEngine {
       resolveSummarize: (params) => this.resolveSummarize(params),
       buildSummarizerLegacyParams: (params) => this.buildSummarizerLegacyParams(params),
       applyAssemblyBudgetCap: (budget) => this.applyAssemblyBudgetCap(budget),
-      refreshBootstrapState: (params) => this.refreshBootstrapState(params),
+      refreshBootstrapState: (params) => this.transcriptReconciler.refreshBootstrapState(params),
       reconcileTranscriptTailForAfterTurn: (params) =>
-        this.reconcileTranscriptTailForAfterTurn(params),
+        this.transcriptReconciler.reconcileTranscriptTailForAfterTurn(params),
       reconcileTranscriptTailForAfterTurnInSessionQueue: (params) =>
-        this.reconcileTranscriptTailForAfterTurnInSessionQueue(params),
+        this.transcriptReconciler.reconcileTranscriptTailForAfterTurnInSessionQueue(params),
     });
 
     this.retrieval = new RetrievalEngine(this.conversationStore, this.summaryStore);
@@ -1635,1916 +1620,6 @@ export class LcmContextEngine implements ContextEngine {
 
   // ── ContextEngine interface ─────────────────────────────────────────────
 
-  private async analyzePersistedTranscriptIdentityOverlaps(params: {
-    conversationId: number;
-    messages: AgentMessage[];
-  }): Promise<{ overlaps: number; firstNonOverlappingIndex: number }> {
-    const existingCounts = new Map<string, number>();
-    const seenCounts = new Map<string, number>();
-    let overlaps = 0;
-    let firstNonOverlappingIndex = -1;
-
-    for (const [index, message] of params.messages.entries()) {
-      const stored = toStoredMessage(message);
-      const identityHash = buildMessageIdentityHash(stored.role, stored.content);
-      const key = `${stored.role}\u0000${identityHash}`;
-      const seen = (seenCounts.get(key) ?? 0) + 1;
-      seenCounts.set(key, seen);
-
-      let existing = existingCounts.get(key);
-      if (existing === undefined) {
-        existing = await this.conversationStore.countMessagesByIdentityHash(
-          params.conversationId,
-          stored.role,
-          identityHash,
-        );
-        existingCounts.set(key, existing);
-      }
-
-      if (seen <= existing) {
-        overlaps += 1;
-      } else if (firstNonOverlappingIndex < 0) {
-        firstNonOverlappingIndex = index;
-      }
-    }
-
-    return { overlaps, firstNonOverlappingIndex };
-  }
-
-  private async countPersistedTranscriptIdentityOverlaps(params: {
-    conversationId: number;
-    messages: AgentMessage[];
-  }): Promise<number> {
-    const analysis = await this.analyzePersistedTranscriptIdentityOverlaps(params);
-    return analysis.overlaps;
-  }
-
-  private async appendOnlyMessagesOverlapPersistedTranscript(params: {
-    conversationId: number;
-    messages: AgentMessage[];
-    /**
-     * The appended slice BEFORE replay/heartbeat filtering. Filtered-out
-     * entries are legitimate parents for surviving entries, so the linear-
-     * chain check consults their ids to avoid false rewrite verdicts.
-     */
-    unfilteredMessages?: AgentMessage[];
-    sessionContext: string;
-    source: string;
-  }): Promise<boolean> {
-    // Entry-id-bearing messages are judged by id, not content identity: a
-    // fresh entry id is a NEW entry even when its content is byte-identical
-    // to persisted history. Repeated identical tool calls are legitimate
-    // (lossless-claw-3071: a tool loop re-reading the same file made the
-    // identity check scream "already persisted" every iteration, forcing a
-    // full re-read per tool call). Two cases still defer to full
-    // reconciliation: an already-persisted entry id (genuine replay), and a
-    // fresh id whose content matches an UNSTAMPED persisted row (flush-lag
-    // catch-up — the full path adopts the id onto that row instead of
-    // importing a duplicate).
-    const idLessMessages: AgentMessage[] = [];
-    let persistedEntryIdOverlaps = 0;
-    let adoptableFlushLagOverlaps = 0;
-    for (const message of params.messages) {
-      const entryId = getTranscriptEntryId(message);
-      if (entryId === null) {
-        idLessMessages.push(message);
-        continue;
-      }
-      if (
-        await this.conversationStore.hasMessageByTranscriptEntryId(
-          params.conversationId,
-          entryId,
-        )
-      ) {
-        persistedEntryIdOverlaps += 1;
-        continue;
-      }
-      const stored = toStoredMessage(message);
-      // Empty stored content (pure tool-call rows) matches trivially against
-      // legacy unstamped rows and proves nothing — and "adopting" onto a
-      // random old empty row would mis-stamp it. Only non-empty identities
-      // indicate a real flush-lagged runtime row.
-      if (stored.content.trim().length === 0) {
-        continue;
-      }
-      if (
-        await this.conversationStore.hasRecentUnstampedMessageByIdentity(
-          params.conversationId,
-          stored.role,
-          stored.content,
-          FLUSH_LAG_ADOPTION_TAIL_WINDOW,
-        )
-      ) {
-        adoptableFlushLagOverlaps += 1;
-      }
-    }
-    if (persistedEntryIdOverlaps > 0 || adoptableFlushLagOverlaps > 0) {
-      this.deps.log.warn(
-        `[lcm] transcript import guard: ${params.source} found ${persistedEntryIdOverlaps} already-persisted transcript entry ids and ${adoptableFlushLagOverlaps} adoptable flush-lagged identities across ${params.messages.length} messages for ${params.sessionContext}; falling back to full reconciliation`,
-      );
-      return true;
-    }
-    // Fresh ids must also EXTEND the persisted tail rather than branch from
-    // an ancestor: a host history rewrite re-issues content under new ids by
-    // reparenting onto an OLDER persisted entry, and those entries need the
-    // full path's stale-id re-stamping, not a blind append. The rewrite
-    // signature is a parent that IS persisted but is NOT the tip and not
-    // part of this appended slice. Parents unknown to the DB are benign —
-    // the append-only checkpoint already verified the file prefix, and
-    // pruned rows, replay-filtered entries, and flush gaps all leave
-    // unknown-parent links that are genuine continuation.
-    if (params.messages.length > idLessMessages.length) {
-      const newestPersistedEntryId = await this.conversationStore.getNewestTranscriptEntryId(
-        params.conversationId,
-      );
-      if (newestPersistedEntryId !== null) {
-        const sliceEntryIds = new Set<string>();
-        for (const message of params.unfilteredMessages ?? params.messages) {
-          const entryId = getTranscriptEntryMeta(message)?.entryId;
-          if (entryId) {
-            sliceEntryIds.add(entryId);
-          }
-        }
-        for (const message of params.messages) {
-          const meta = getTranscriptEntryMeta(message);
-          const parentId = meta?.parentId ?? null;
-          if (
-            parentId === null ||
-            parentId === newestPersistedEntryId ||
-            sliceEntryIds.has(parentId)
-          ) {
-            continue;
-          }
-          if (
-            await this.conversationStore.hasMessageByTranscriptEntryId(
-              params.conversationId,
-              parentId,
-            )
-          ) {
-            this.deps.log.warn(
-              `[lcm] transcript import guard: ${params.source} appended entry reparents onto a non-tip persisted entry (suffix rewrite) for ${params.sessionContext}; falling back to full reconciliation`,
-            );
-            return true;
-          }
-        }
-      }
-    }
-    if (idLessMessages.length === 0) {
-      return false;
-    }
-
-    const overlaps = await this.countPersistedTranscriptIdentityOverlaps({
-      conversationId: params.conversationId,
-      messages: idLessMessages,
-    });
-    if (overlaps === 0) {
-      return false;
-    }
-
-    this.deps.log.warn(
-      `[lcm] transcript import guard: ${params.source} found ${overlaps}/${idLessMessages.length} already-persisted message identities (id-less entries) for ${params.sessionContext}; falling back to full reconciliation`,
-    );
-    return true;
-  }
-
-  /**
-   * Reconcile session-file history with persisted messages and append only the
-   * tail that is present in JSONL but missing from LCM.
-   */
-  /**
-   * Exact reconciliation for transcripts whose entries all carry stable
-   * envelope ids: anchor on the checkpoint's last processed entry id (or the
-   * newest id already persisted), then import only the tail entries whose
-   * ids are missing — adopting identity-matched rows that were persisted
-   * without an id (runtime flush lag, pre-migration data) instead of
-   * importing duplicates. Entry-id matching is immune to content rewriting
-   * (externalized tool results), which defeats content-identity anchors.
-   *
-   * Returns null when no id lineage links the transcript to this
-   * conversation; the caller's content-identity machinery and no-anchor
-   * guards then decide.
-   */
-  private async reconcileSessionTailByEntryIds(params: {
-    sessionId: string;
-    sessionKey?: string;
-    conversationId: number;
-    historicalMessages: AgentMessage[];
-    entryIds: string[];
-    lastProcessedEntryId?: string | null;
-    existingDbCount: number;
-    sessionContext: string;
-    startedAt: number;
-  }): Promise<TranscriptReconcileResult | null> {
-    const { conversationId, historicalMessages, entryIds, sessionContext, startedAt } = params;
-
-    // Query existence only for the tail when the checkpoint anchor is still
-    // in the transcript; otherwise probe every id to find the newest
-    // persisted one.
-    const checkpointAnchorIndex = params.lastProcessedEntryId
-      ? entryIds.lastIndexOf(params.lastProcessedEntryId)
-      : -1;
-    const knownExisting = await this.conversationStore.filterExistingTranscriptEntryIds(
-      conversationId,
-      checkpointAnchorIndex >= 0 ? entryIds.slice(checkpointAnchorIndex + 1) : entryIds,
-    );
-    const selection = selectEntryIdTail({
-      entryIds,
-      existingEntryIds: knownExisting,
-      lastProcessedEntryId: params.lastProcessedEntryId,
-    });
-
-    if (selection.kind === "no-id-lineage") {
-      return null;
-    }
-    if (selection.kind === "at-tip") {
-      this.deps.log.debug(
-        `[lcm] reconcileSessionTail: entry-id anchor at tip for ${sessionContext} duration=${formatDurationMs(Date.now() - startedAt)} historicalMessages=${historicalMessages.length} importedMessages=0 overlap=true`,
-      );
-      return { blockedByImportCap: false, importedMessages: 0, hasOverlap: true };
-    }
-
-    const anchorIndex = selection.anchorIndex;
-    const candidates = selection.missingIndexes.map((index) => historicalMessages[index]!);
-    const missingTail = this.filterSyntheticHeartbeatTranscriptMessages({
-      messages: candidates,
-      sessionContext,
-      source: "reconcileSessionTail entry-id",
-    });
-
-    // Entry-id anchored backlogs are exact — lineage is proven by the id
-    // anchor and every import is individually id-verified — so a backlog
-    // beyond the cap drains in bounded oldest-first chunks instead of
-    // freezing, mirroring the anchored content path. The checkpoint does
-    // not advance while a pass is capped, so repeated passes converge.
-    const entryIdImportCap = transcriptImportCap(params.existingDbCount);
-    const entryIdImportCapped =
-      params.existingDbCount > 0 && missingTail.length > entryIdImportCap;
-    const importableTail = entryIdImportCapped
-      ? missingTail.slice(0, entryIdImportCap)
-      : missingTail;
-    if (entryIdImportCapped) {
-      this.deps.log.warn(
-        `[lcm] reconcileSessionTail: entry-id import cap chunking for ${sessionContext} — importing ${importableTail.length}/${missingTail.length} anchored backlog messages this pass (existing: ${params.existingDbCount}, cap: ${entryIdImportCap}); remaining backlog continues next pass`,
-      );
-    }
-
-    // Ids on the current leaf path. A persisted row whose entry id is NOT in
-    // this set was stranded by a host history rewrite (the suffix was
-    // re-appended under new ids) and is eligible for stale-id re-stamping.
-    const leafEntryIds = new Set(entryIds);
-
-    let importedMessages = 0;
-    let adoptedMessages = 0;
-    let restampedMessages = 0;
-    for (const message of importableTail) {
-      const entryId = getTranscriptEntryId(message)!;
-      const stored = toStoredMessage(message);
-      const adopted = await this.conversationStore.adoptTranscriptEntryId(
-        conversationId,
-        stored.role,
-        stored.content,
-        entryId,
-      );
-      if (adopted) {
-        adoptedMessages += 1;
-        continue;
-      }
-      const restamped = await this.adoptStaleTranscriptEntryId({
-        conversationId,
-        leafEntryIds,
-        role: stored.role,
-        content: stored.content,
-        entryId,
-      });
-      if (restamped) {
-        restampedMessages += 1;
-        continue;
-      }
-      // Entry-id-verified imports are exact (the id is proven absent), so the
-      // same-second replay flood heuristic does not apply.
-      const result = await this.ingestSingle({
-        sessionId: params.sessionId,
-        sessionKey: params.sessionKey,
-        message,
-        skipReplayTimestampFloodGuard: true,
-      });
-      if (result.ingested) {
-        importedMessages += 1;
-      }
-    }
-
-    this.deps.log.debug(
-      `[lcm] reconcileSessionTail: entry-id path for ${sessionContext} duration=${formatDurationMs(Date.now() - startedAt)} historicalMessages=${historicalMessages.length} anchorIndex=${anchorIndex} missingTail=${missingTail.length} importedMessages=${importedMessages} adoptedMessages=${adoptedMessages} restampedMessages=${restampedMessages} capped=${entryIdImportCapped}`,
-    );
-    if (entryIdImportCapped) {
-      return {
-        blockedByImportCap: true,
-        blockedReason: "import-cap",
-        importedMessages,
-        hasOverlap: true,
-      };
-    }
-    return { blockedByImportCap: false, importedMessages, hasOverlap: true };
-  }
-
-  /**
-   * Re-stamp an identity-matched row whose stored entry id has left the
-   * transcript's leaf path. Host history rewrites (rewriteTranscriptEntries,
-   * the host's own tool-result truncation, gateway chat edits) re-append the
-   * active suffix under freshly generated ids; the stranded rows are the
-   * same messages, so they adopt the re-issued id instead of importing a
-   * duplicate. Rows whose ids are still on the leaf path are live entries
-   * and never touched. Returns true when a row was re-stamped.
-   */
-  private async adoptStaleTranscriptEntryId(params: {
-    conversationId: number;
-    leafEntryIds: ReadonlySet<string>;
-    role: StoredMessage["role"];
-    content: string;
-    entryId: string;
-  }): Promise<boolean> {
-    const candidates = await this.conversationStore.listTranscriptEntryIdsByIdentity(
-      params.conversationId,
-      params.role,
-      params.content,
-    );
-    for (const candidate of candidates) {
-      if (params.leafEntryIds.has(candidate.transcriptEntryId)) {
-        continue;
-      }
-      const restamped = await this.conversationStore.restampTranscriptEntryId(
-        candidate.messageId,
-        params.entryId,
-      );
-      if (restamped) {
-        return true;
-      }
-    }
-    return false;
-  }
-
-  private async reconcileSessionTail(params: {
-    sessionId: string;
-    sessionKey?: string;
-    conversationId: number;
-    historicalMessages: AgentMessage[];
-    checkpointEntryHash?: string | null;
-    lastProcessedEntryId?: string | null;
-    skipContentAnchorScan?: boolean;
-    allowNoAnchorImport?: boolean;
-    noAnchorImportReason?: string;
-  }): Promise<TranscriptReconcileResult> {
-    const { sessionId, conversationId, historicalMessages } = params;
-    const startedAt = Date.now();
-    const sessionContext = this.formatSessionLogContext({
-      conversationId,
-      sessionId,
-      sessionKey: params.sessionKey,
-    });
-    if (historicalMessages.length === 0) {
-      this.deps.log.debug(
-        `[lcm] reconcileSessionTail: skipped for ${sessionContext} duration=${formatDurationMs(Date.now() - startedAt)} historicalMessages=0 reason=empty-history`,
-      );
-      return { blockedByImportCap: false, importedMessages: 0, hasOverlap: false };
-    }
-
-    const latestDbMessage = await this.conversationStore.getLastMessage(conversationId);
-    if (!latestDbMessage) {
-      this.deps.log.debug(
-        `[lcm] reconcileSessionTail: skipped for ${sessionContext} duration=${formatDurationMs(Date.now() - startedAt)} historicalMessages=${historicalMessages.length} reason=no-db-tail`,
-      );
-      return { blockedByImportCap: false, importedMessages: 0, hasOverlap: false };
-    }
-    const existingDbCount = await this.conversationStore.getMessageCount(conversationId);
-
-    // Exact path: when every transcript entry carries a stable envelope id,
-    // anchor and diff by id instead of content identity.
-    const candidateEntryIds = historicalMessages.map((message) => getTranscriptEntryId(message));
-    if (candidateEntryIds.every((entryId): entryId is string => entryId !== null)) {
-      const entryIdResult = await this.reconcileSessionTailByEntryIds({
-        sessionId,
-        sessionKey: params.sessionKey,
-        conversationId,
-        historicalMessages,
-        entryIds: candidateEntryIds,
-        lastProcessedEntryId: params.lastProcessedEntryId,
-        existingDbCount,
-        sessionContext,
-        startedAt,
-      });
-      if (entryIdResult) {
-        return entryIdResult;
-      }
-    }
-
-    const storedHistoricalMessages = historicalMessages.map((message) => toStoredMessage(message));
-
-    // Fast path: one tail comparison for the common in-sync case.
-    const latestHistorical = storedHistoricalMessages[storedHistoricalMessages.length - 1];
-    const latestIdentity = messageIdentity(latestDbMessage.role, latestDbMessage.content);
-    if (
-      !params.skipContentAnchorScan &&
-      latestIdentity === messageIdentity(latestHistorical.role, latestHistorical.content)
-    ) {
-      const dbOccurrences = await this.conversationStore.countMessagesByIdentity(
-        conversationId,
-        latestDbMessage.role,
-        latestDbMessage.content,
-      );
-      let historicalOccurrences = 0;
-      for (const stored of storedHistoricalMessages) {
-        if (messageIdentity(stored.role, stored.content) === latestIdentity) {
-          historicalOccurrences += 1;
-        }
-      }
-      if (dbOccurrences === historicalOccurrences) {
-        this.deps.log.debug(
-          `[lcm] reconcileSessionTail: fast path for ${sessionContext} duration=${formatDurationMs(Date.now() - startedAt)} historicalMessages=${historicalMessages.length} importedMessages=0 overlap=true`,
-        );
-        return { blockedByImportCap: false, importedMessages: 0, hasOverlap: true };
-      }
-    }
-
-    // Slow path: walk backward through JSONL to find the most recent anchor
-    // message that already exists in LCM, then append everything after it.
-    let anchorIndex = -1;
-    const historicalIdentityTotals = new Map<string, number>();
-    for (const stored of storedHistoricalMessages) {
-      const identity = messageIdentity(stored.role, stored.content);
-      historicalIdentityTotals.set(identity, (historicalIdentityTotals.get(identity) ?? 0) + 1);
-    }
-
-    if (!params.skipContentAnchorScan) {
-      const historicalIdentityCountsAfterIndex = new Map<string, number>();
-      const dbIdentityCounts = new Map<string, number>();
-      for (let index = storedHistoricalMessages.length - 1; index >= 0; index--) {
-        const stored = storedHistoricalMessages[index];
-        const identity = messageIdentity(stored.role, stored.content);
-        const seenAfter = historicalIdentityCountsAfterIndex.get(identity) ?? 0;
-        const total = historicalIdentityTotals.get(identity) ?? 0;
-        const occurrencesThroughIndex = total - seenAfter;
-        const exists = await this.conversationStore.hasMessage(
-          conversationId,
-          stored.role,
-          stored.content,
-        );
-        historicalIdentityCountsAfterIndex.set(identity, seenAfter + 1);
-        if (!exists) {
-          continue;
-        }
-
-        let dbCountForIdentity = dbIdentityCounts.get(identity);
-        if (dbCountForIdentity === undefined) {
-          dbCountForIdentity = await this.conversationStore.countMessagesByIdentity(
-            conversationId,
-            stored.role,
-            stored.content,
-          );
-          dbIdentityCounts.set(identity, dbCountForIdentity);
-        }
-
-        // Match the same occurrence index as the DB tail so repeated empty
-        // tool messages do not anchor against a later, still-missing entry.
-        if (dbCountForIdentity !== occurrencesThroughIndex) {
-          continue;
-        }
-
-        anchorIndex = index;
-        break;
-      }
-    }
-
-    if (anchorIndex < 0) {
-      const checkpointEntryHash = params.checkpointEntryHash;
-      if (checkpointEntryHash) {
-        // Externalized bootstrap rows no longer match raw JSONL content, so
-        // fall back to the raw transcript checkpoint before declaring no overlap.
-        for (let index = storedHistoricalMessages.length - 1; index >= 0; index--) {
-          if (createBootstrapEntryHash(storedHistoricalMessages[index]) === checkpointEntryHash) {
-            anchorIndex = index;
-            break;
-          }
-        }
-      }
-
-      if (anchorIndex < 0) {
-        if (params.allowNoAnchorImport) {
-          if (
-            (params.noAnchorImportReason === "path-mismatch" ||
-              params.noAnchorImportReason === "checkpoint-missing-recovery") &&
-            isLikelyInjectedDeliveryOnlyTranscript(historicalMessages)
-          ) {
-            this.deps.log.warn(
-              `[lcm] reconcileSessionTail: blocked delivery-only path-mismatched transcript for ${sessionContext}; preserving existing checkpoint because the rotated transcript contains only injected delivery/config traffic`,
-            );
-            this.deps.log.debug(
-              `[lcm] reconcileSessionTail: blocked delivery-only path mismatch for ${sessionContext} duration=${formatDurationMs(Date.now() - startedAt)} historicalMessages=${historicalMessages.length} overlap=false`,
-            );
-            return { blockedByImportCap: false, importedMessages: 0, hasOverlap: false };
-          }
-
-          const replayAnalysis = await this.analyzePersistedTranscriptIdentityOverlaps({
-            conversationId,
-            messages: historicalMessages,
-          });
-          const persistedIdentityOverlaps = replayAnalysis.overlaps;
-          let noAnchorImportMessages = this.filterSyntheticHeartbeatTranscriptMessages({
-            messages: historicalMessages,
-            sessionContext,
-            source: "reconcileSessionTail no-anchor",
-          });
-          // A fully id-bearing batch resolves identity overlaps exactly:
-          // identity matches adopt the re-issued entry id (host rewrites and
-          // rotations re-append the surviving suffix under new ids) and only
-          // genuinely-new entries import. The replay-overlap block below
-          // exists for id-less ambiguity, where overlap could equally mean a
-          // replayed file; applying it here would freeze rewritten-epoch
-          // transcripts instead of healing them.
-          const allEntryIdBearing =
-            noAnchorImportMessages.length > 0 &&
-            noAnchorImportMessages.every((message) => getTranscriptEntryId(message) !== null);
-          const replayThreshold = Math.max(3, Math.ceil(historicalMessages.length * 0.5));
-          if (!allEntryIdBearing && persistedIdentityOverlaps >= replayThreshold) {
-            if (replayAnalysis.firstNonOverlappingIndex < 0) {
-              this.deps.log.warn(
-                `[lcm] reconcileSessionTail: duplicate transcript replay blocked for ${sessionContext} - ${persistedIdentityOverlaps}/${historicalMessages.length} candidate messages already exist (reason: ${params.noAnchorImportReason ?? "unspecified"}). Aborting to prevent replay flood.`,
-              );
-              this.deps.log.debug(
-                `[lcm] reconcileSessionTail: blocked duplicate transcript replay for ${sessionContext} duration=${formatDurationMs(Date.now() - startedAt)} historicalMessages=${historicalMessages.length} persistedIdentityOverlaps=${persistedIdentityOverlaps} overlap=false`,
-              );
-              return {
-                blockedByImportCap: true,
-                blockedReason: "duplicate-transcript-replay",
-                importedMessages: 0,
-                hasOverlap: false,
-              };
-            }
-
-            if (replayAnalysis.firstNonOverlappingIndex > 0) {
-              noAnchorImportMessages = this.filterSyntheticHeartbeatTranscriptMessages({
-                messages: historicalMessages.slice(replayAnalysis.firstNonOverlappingIndex),
-                sessionContext,
-                source: "reconcileSessionTail no-anchor replay-prefix",
-              });
-              this.deps.log.warn(
-                `[lcm] reconcileSessionTail: duplicate transcript replay guard dropped ${replayAnalysis.firstNonOverlappingIndex}/${historicalMessages.length} already-persisted prefix messages for ${sessionContext} before no-anchor import (reason: ${params.noAnchorImportReason ?? "unspecified"})`,
-              );
-            }
-          }
-
-          const importCap = transcriptImportCap(existingDbCount);
-          let noAnchorImportCapped = false;
-          if (noAnchorImportMessages.length > importCap) {
-            if (allEntryIdBearing) {
-              // A fully id-bearing new epoch is exact (every entry adopts or
-              // imports by verified id), so a large rollover — e.g. a host
-              // rewrite or compaction successor with a kept tail beyond the
-              // cap — drains in bounded oldest-first chunks instead of
-              // freezing before adoption can heal it. After the first chunk
-              // persists ids, the entry-id anchor takes over on later passes.
-              noAnchorImportCapped = true;
-              this.deps.log.warn(
-                `[lcm] reconcileSessionTail: no-anchor entry-id import cap chunking for ${sessionContext} — importing ${importCap}/${noAnchorImportMessages.length} new-epoch messages this pass (existing: ${existingDbCount}, cap: ${importCap}, reason: ${params.noAnchorImportReason ?? "unspecified"}); remaining backlog continues next pass`,
-              );
-              noAnchorImportMessages = noAnchorImportMessages.slice(0, importCap);
-            } else {
-              this.deps.log.warn(
-                `[lcm] reconcileSessionTail: no anchor import cap exceeded for ${sessionContext} - would import ${noAnchorImportMessages.length} messages (existing: ${existingDbCount}, cap: ${importCap}, reason: ${params.noAnchorImportReason ?? "unspecified"}). Aborting to prevent flood.`,
-              );
-              this.deps.log.debug(
-                `[lcm] reconcileSessionTail: blocked no-anchor import for ${sessionContext} duration=${formatDurationMs(Date.now() - startedAt)} historicalMessages=${historicalMessages.length} candidateMessages=${noAnchorImportMessages.length} existingDbCount=${existingDbCount} cap=${importCap} overlap=false`,
-              );
-              return {
-                blockedByImportCap: true,
-                blockedReason: "import-cap",
-                importedMessages: 0,
-                hasOverlap: false,
-              };
-            }
-          }
-
-          if (params.noAnchorImportReason === "same-path-shrink") {
-            const rawIdMatches = this.countActiveCrossConversationRawIdMatches({
-              conversationId,
-              sessionId,
-              messages: noAnchorImportMessages,
-            });
-            if (rawIdMatches.matchedRawIds > 0) {
-              this.deps.log.warn(
-                `[lcm] reconcileSessionTail: blocked same-path-shrink no-anchor import for ${sessionContext} because ${rawIdMatches.matchedRawIds}/${rawIdMatches.candidateRawIds} candidate raw ids already exist in other active conversations`,
-              );
-              this.deps.log.debug(
-                `[lcm] reconcileSessionTail: blocked cross-conversation raw-id duplicate for ${sessionContext} duration=${formatDurationMs(Date.now() - startedAt)} historicalMessages=${historicalMessages.length} candidateRawIds=${rawIdMatches.candidateRawIds} matchedRawIds=${rawIdMatches.matchedRawIds} overlap=false`,
-              );
-              return {
-                blockedByImportCap: true,
-                blockedReason: "cross-conversation-raw-id",
-                importedMessages: 0,
-                hasOverlap: false,
-              };
-            }
-          }
-
-          // Ids on the current leaf path, for stale-id re-stamping of rows
-          // stranded by the rewrite/rotation that produced this new epoch.
-          const noAnchorLeafEntryIds = new Set(
-            historicalMessages
-              .map((message) => getTranscriptEntryId(message))
-              .filter((id): id is string => id !== null),
-          );
-          let importedMessages = 0;
-          let adoptedMessages = 0;
-          for (const message of noAnchorImportMessages) {
-            const entryId = getTranscriptEntryId(message);
-            if (entryId) {
-              const alreadyPersisted = await this.conversationStore.hasMessageByTranscriptEntryId(
-                conversationId,
-                entryId,
-              );
-              if (alreadyPersisted) {
-                continue;
-              }
-              const stored = toStoredMessage(message);
-              const adopted =
-                (await this.conversationStore.adoptTranscriptEntryId(
-                  conversationId,
-                  stored.role,
-                  stored.content,
-                  entryId,
-                )) ||
-                (await this.adoptStaleTranscriptEntryId({
-                  conversationId,
-                  leafEntryIds: noAnchorLeafEntryIds,
-                  role: stored.role,
-                  content: stored.content,
-                  entryId,
-                }));
-              if (adopted) {
-                adoptedMessages += 1;
-                continue;
-              }
-            }
-            const result = await this.ingestSingle({
-              sessionId,
-              sessionKey: params.sessionKey,
-              message,
-              skipReplayTimestampFloodGuard: true,
-            });
-            if (result.ingested) {
-              importedMessages += 1;
-            }
-          }
-          this.deps.log.warn(
-            `[lcm] reconcileSessionTail: no anchor for ${sessionContext}; imported transcript as new epoch reason=${params.noAnchorImportReason ?? "unspecified"} duration=${formatDurationMs(Date.now() - startedAt)} historicalMessages=${historicalMessages.length} candidateMessages=${noAnchorImportMessages.length} importedMessages=${importedMessages} adoptedMessages=${adoptedMessages} capped=${noAnchorImportCapped} overlap=${adoptedMessages > 0}`,
-          );
-          if (noAnchorImportCapped) {
-            // Partial pass: keep the checkpoint un-advanced so the next pass
-            // continues the drain (via the entry-id anchor once this chunk's
-            // ids are persisted).
-            return {
-              blockedByImportCap: true,
-              blockedReason: "import-cap",
-              importedMessages,
-              hasOverlap: adoptedMessages > 0,
-            };
-          }
-          // Adoption proves the "new epoch" overlaps persisted history (the
-          // host re-issued ids for messages we already hold), so report the
-          // overlap — the caller then refreshes the checkpoint instead of
-          // re-entering this path every turn.
-          return { blockedByImportCap: false, importedMessages, hasOverlap: adoptedMessages > 0 };
-        }
-        this.deps.log.debug(
-          `[lcm] reconcileSessionTail: no anchor for ${sessionContext} duration=${formatDurationMs(Date.now() - startedAt)} historicalMessages=${historicalMessages.length} importedMessages=0 overlap=false`,
-        );
-        return { blockedByImportCap: false, importedMessages: 0, hasOverlap: false };
-      }
-    }
-    if (anchorIndex >= historicalMessages.length - 1) {
-      this.deps.log.debug(
-        `[lcm] reconcileSessionTail: anchor at tip for ${sessionContext} duration=${formatDurationMs(Date.now() - startedAt)} historicalMessages=${historicalMessages.length} importedMessages=0 overlap=true`,
-      );
-      return { blockedByImportCap: false, importedMessages: 0, hasOverlap: true };
-    }
-
-    const missingTailFiltered = await this.filterBootstrapReplayMessages({
-      messages: historicalMessages.slice(anchorIndex + 1),
-      sessionContext,
-      source: "reconcileSessionTail",
-      priorMessages: historicalMessages.slice(0, anchorIndex + 1),
-    });
-    const missingTail = this.filterSyntheticHeartbeatTranscriptMessages({
-      messages: missingTailFiltered.messages,
-      sessionContext,
-      source: "reconcileSessionTail",
-    });
-
-    // Anchored missing tails are this conversation's own continuation
-    // (lineage proven by the identity anchor), so a tail larger than the
-    // cap must not freeze reconcile forever — that wedges afterTurn
-    // persistence while the backlog keeps growing. Import a capped
-    // oldest-first chunk per pass instead: order is preserved, per-pass
-    // flood exposure stays bounded, and the growing existing count raises
-    // the cap so repeated passes converge. The checkpoint/frontier still
-    // does not advance until the backlog fully drains (blockedByImportCap
-    // stays true on partial passes).
-    const anchoredImportCap = transcriptImportCap(existingDbCount);
-    const importCapped = existingDbCount > 0 && missingTail.length > anchoredImportCap;
-    const importableTail = importCapped ? missingTail.slice(0, anchoredImportCap) : missingTail;
-    if (importCapped) {
-      this.deps.log.warn(
-        `[lcm] reconcileSessionTail: import cap chunking for ${sessionContext} — importing ${importableTail.length}/${missingTail.length} anchored backlog messages this pass (existing: ${existingDbCount}, cap: ${anchoredImportCap}); remaining backlog continues next pass`,
-      );
-    }
-
-    let importedMessages = 0;
-    for (const [index, message] of importableTail.entries()) {
-      const result = await this.ingestSingle({
-        sessionId,
-        sessionKey: params.sessionKey,
-        message,
-        skipReplayTimestampFloodGuard:
-          index < missingTailFiltered.replayGuardExemptPrefixLength,
-      });
-      if (result.ingested) {
-        importedMessages += 1;
-      }
-    }
-
-    if (importCapped) {
-      this.deps.log.debug(
-        `[lcm] reconcileSessionTail: capped chunk for ${sessionContext} duration=${formatDurationMs(Date.now() - startedAt)} historicalMessages=${historicalMessages.length} anchorIndex=${anchorIndex} missingTail=${missingTail.length} importedMessages=${importedMessages} existingDbCount=${existingDbCount} cap=${anchoredImportCap}`,
-      );
-      return {
-        blockedByImportCap: true,
-        blockedReason: "import-cap",
-        importedMessages,
-        hasOverlap: true,
-      };
-    }
-
-    this.deps.log.debug(
-      `[lcm] reconcileSessionTail: slow path for ${sessionContext} duration=${formatDurationMs(Date.now() - startedAt)} historicalMessages=${historicalMessages.length} anchorIndex=${anchorIndex} missingTail=${missingTail.length} importedMessages=${importedMessages}`,
-    );
-    return { blockedByImportCap: false, importedMessages, hasOverlap: true };
-  }
-
-  /** Count candidate raw event IDs that already belong to another active conversation. */
-  private countActiveCrossConversationRawIdMatches(params: {
-    conversationId: number;
-    sessionId: string;
-    messages: AgentMessage[];
-  }): { candidateRawIds: number; matchedRawIds: number } {
-    const candidateRawIds = new Set<string>();
-    for (const message of params.messages) {
-      const stored = toStoredMessage(message);
-      const parts = buildMessageParts({
-        sessionId: params.sessionId,
-        message,
-        fallbackContent: stored.content,
-      });
-      for (const part of parts) {
-        for (const rawId of extractRawIdsFromPartMetadata(part.metadata)) {
-          candidateRawIds.add(rawId);
-        }
-      }
-    }
-
-    if (candidateRawIds.size === 0) {
-      return { candidateRawIds: 0, matchedRawIds: 0 };
-    }
-
-    const matchStmt = this.db.prepare(
-      `SELECT 1 AS found
-       FROM message_parts mp
-       JOIN messages m ON m.message_id = mp.message_id
-       JOIN conversations c ON c.conversation_id = m.conversation_id
-       WHERE c.active = 1
-         AND m.conversation_id <> ?
-         AND mp.metadata IS NOT NULL
-         AND json_valid(mp.metadata)
-         AND (
-           json_extract(mp.metadata, '$.raw.id') = ?
-           OR json_extract(mp.metadata, '$.raw.call_id') = ?
-           OR json_extract(mp.metadata, '$.raw.toolCallId') = ?
-           OR json_extract(mp.metadata, '$.raw.tool_call_id') = ?
-           OR json_extract(mp.metadata, '$.raw.toolUseId') = ?
-           OR json_extract(mp.metadata, '$.raw.tool_use_id') = ?
-           OR mp.tool_call_id = ?
-           OR json_extract(mp.metadata, '$.id') = ?
-           OR json_extract(mp.metadata, '$.call_id') = ?
-           OR json_extract(mp.metadata, '$.toolCallId') = ?
-           OR json_extract(mp.metadata, '$.tool_call_id') = ?
-           OR json_extract(mp.metadata, '$.toolUseId') = ?
-           OR json_extract(mp.metadata, '$.tool_use_id') = ?
-         )
-       LIMIT 1`,
-    );
-
-    let matchedRawIds = 0;
-    for (const rawId of candidateRawIds) {
-      const row = matchStmt.get(
-        params.conversationId,
-        rawId,
-        rawId,
-        rawId,
-        rawId,
-        rawId,
-        rawId,
-        rawId,
-        rawId,
-        rawId,
-        rawId,
-        rawId,
-        rawId,
-        rawId,
-      ) as { found: number } | undefined;
-      if (row?.found === 1) {
-        matchedRawIds += 1;
-      }
-    }
-
-    return { candidateRawIds: candidateRawIds.size, matchedRawIds };
-  }
-
-  /** Drop exact raw-id transcript replays while preserving content-only repeated user turns. */
-  private async filterPersistedRawIdReplayBatch(params: {
-    conversationId: number;
-    sessionId: string;
-    sessionKey?: string;
-    messages: AgentMessage[];
-  }): Promise<AgentMessage[]> {
-    const idMatchPredicate = `(
-      json_extract(mp.metadata, '$.raw.id') = ?
-      OR json_extract(mp.metadata, '$.raw.call_id') = ?
-      OR json_extract(mp.metadata, '$.raw.toolCallId') = ?
-      OR json_extract(mp.metadata, '$.raw.tool_call_id') = ?
-      OR json_extract(mp.metadata, '$.raw.toolUseId') = ?
-      OR json_extract(mp.metadata, '$.raw.tool_use_id') = ?
-      OR mp.tool_call_id = ?
-      OR json_extract(mp.metadata, '$.id') = ?
-      OR json_extract(mp.metadata, '$.call_id') = ?
-      OR json_extract(mp.metadata, '$.toolCallId') = ?
-      OR json_extract(mp.metadata, '$.tool_call_id') = ?
-      OR json_extract(mp.metadata, '$.toolUseId') = ?
-      OR json_extract(mp.metadata, '$.tool_use_id') = ?
-    )`;
-    const rawCoverageStmt = this.db.prepare(
-      `SELECT m.message_id AS messageId
-       FROM message_parts mp
-       JOIN messages m ON m.message_id = mp.message_id
-       WHERE m.conversation_id = ?
-         AND m.role = ?
-         AND mp.metadata IS NOT NULL
-         AND json_valid(mp.metadata)
-         AND ${idMatchPredicate}`,
-    );
-    const identityCoverageStmt = this.db.prepare(
-      `SELECT m.message_id AS messageId
-       FROM message_parts mp
-       JOIN messages m ON m.message_id = mp.message_id
-       WHERE m.conversation_id = ?
-         AND m.role = ?
-         AND m.identity_hash = ?
-         AND mp.metadata IS NOT NULL
-         AND json_valid(mp.metadata)
-         AND ${idMatchPredicate}`,
-    );
-    const externalizedCoverageStmt = this.db.prepare(
-      `SELECT
-         m.message_id AS messageId,
-         json_extract(mp.metadata, '$.externalizedFileId') AS fileId,
-         json_extract(mp.metadata, '$.originalByteSize') AS originalByteSize,
-         mp.metadata AS metadata
-       FROM message_parts mp
-       JOIN messages m ON m.message_id = mp.message_id
-       WHERE m.conversation_id = ?
-         AND m.role = ?
-         AND mp.metadata IS NOT NULL
-         AND json_valid(mp.metadata)
-         AND json_extract(mp.metadata, '$.externalizationReason') = 'large_tool_result'
-         AND ${idMatchPredicate}`,
-    );
-    const rawBlockSignatureStmt = this.db.prepare(
-      `SELECT metadata
-       FROM message_parts
-       WHERE message_id = ?
-       ORDER BY ordinal ASC`,
-    );
-
-    const filtered: AgentMessage[] = [];
-    let replayedMessages = 0;
-
-    for (const message of params.messages) {
-      const stored = toStoredMessage(message);
-      const replayIds = new Set<string>();
-      const rawBlockIds = new Set<string>();
-      const rawBlockSignatures: string[] = [];
-      const replayIdsByPart: string[][] = [];
-      let everyPartHasRawBlockId = true;
-      const parts = buildMessageParts({
-        sessionId: params.sessionId,
-        message,
-        fallbackContent: stored.content,
-      });
-      for (const part of parts) {
-        const partRawBlockIds = extractRawBlockIdsFromPartMetadata(part.metadata);
-        if (partRawBlockIds.length === 0) {
-          everyPartHasRawBlockId = false;
-        }
-        for (const rawId of partRawBlockIds) {
-          rawBlockIds.add(rawId);
-        }
-        const rawBlockSignature = extractRawBlockSignatureFromPartMetadata(part.metadata);
-        if (rawBlockSignature) {
-          rawBlockSignatures.push(rawBlockSignature);
-        }
-        const partReplayIds = extractRawIdsFromPartMetadata(part.metadata);
-        replayIdsByPart.push(partReplayIds);
-        for (const rawId of partReplayIds) {
-          replayIds.add(rawId);
-        }
-      }
-
-      if (replayIds.size === 0) {
-        filtered.push(message);
-        continue;
-      }
-
-      const canMatchWithoutIdentity = rawBlockIds.size > 0 && everyPartHasRawBlockId;
-      const matchedIds = canMatchWithoutIdentity ? rawBlockIds : replayIds;
-      const externalizedTextsById = extractPlainToolReplayTextsById(message);
-      const coverageByMessageId = new Map<number, Set<string>>();
-      for (const rawId of matchedIds) {
-        const rawIdArgs = [
-          rawId,
-          rawId,
-          rawId,
-          rawId,
-          rawId,
-          rawId,
-          rawId,
-          rawId,
-          rawId,
-          rawId,
-          rawId,
-          rawId,
-          rawId,
-        ];
-        let rows: Array<{ messageId: number }>;
-        if (canMatchWithoutIdentity) {
-          rows = rawCoverageStmt.all(
-            params.conversationId,
-            stored.role,
-            ...rawIdArgs,
-          ) as Array<{ messageId: number }>;
-        } else {
-          const identityHash = buildMessageIdentityHash(stored.role, stored.content);
-          rows = identityCoverageStmt.all(
-            params.conversationId,
-            stored.role,
-            identityHash,
-            ...rawIdArgs,
-          ) as Array<{ messageId: number }>;
-        }
-        for (const row of rows) {
-          const matchedRawIds = coverageByMessageId.get(row.messageId) ?? new Set<string>();
-          matchedRawIds.add(rawId);
-          coverageByMessageId.set(row.messageId, matchedRawIds);
-        }
-      }
-
-      let alreadyPersisted = false;
-      if (canMatchWithoutIdentity) {
-        for (const [messageId, rawIds] of coverageByMessageId.entries()) {
-          if (rawIds.size !== matchedIds.size) {
-            continue;
-          }
-          const rows = rawBlockSignatureStmt.all(messageId) as Array<{ metadata: string | null }>;
-          if (rows.length !== parts.length) {
-            continue;
-          }
-          let allPartsMatch = true;
-          for (let index = 0; index < rows.length; index += 1) {
-            const persistedMetadata = rows[index]!.metadata;
-            const persistedSignature = extractRawBlockSignatureFromPartMetadata(persistedMetadata);
-            if (
-              persistedSignature === rawBlockSignatures[index] &&
-              externalizedReplayMetadataMatches(persistedMetadata, parts[index]?.metadata)
-            ) {
-              continue;
-            }
-            let externalizedPartMatches = false;
-            for (const rawId of replayIdsByPart[index] ?? []) {
-              if (!extractRawIdsFromPartMetadata(persistedMetadata).includes(rawId)) {
-                continue;
-              }
-              const externalizedText = externalizedTextsById.get(rawId);
-              if (externalizedText === undefined) {
-                continue;
-              }
-              let persistedParsed: unknown;
-              try {
-                persistedParsed = persistedMetadata ? JSON.parse(persistedMetadata) : undefined;
-              } catch {
-                continue;
-              }
-              const persistedRecord = asRecord(persistedParsed);
-              const fileId = safeString(persistedRecord?.externalizedFileId);
-              const originalByteSize = persistedRecord?.originalByteSize;
-              if (
-                !fileId ||
-                Number(originalByteSize) !== Buffer.byteLength(externalizedText, "utf8")
-              ) {
-                continue;
-              }
-              const largeFile = await this.summaryStore.getLargeFile(fileId);
-              if (!largeFile) {
-                continue;
-              }
-              let storedText: string;
-              try {
-                storedText = readFileSync(largeFile.storageUri, "utf8");
-              } catch {
-                continue;
-              }
-              if (
-                storedText === externalizedText &&
-                externalizedReplayMetadataMatches(persistedMetadata, parts[index]?.metadata)
-              ) {
-                externalizedPartMatches = true;
-                break;
-              }
-            }
-            if (!externalizedPartMatches) {
-              allPartsMatch = false;
-              break;
-            }
-          }
-          if (allPartsMatch) {
-            alreadyPersisted = true;
-            break;
-          }
-        }
-      } else {
-        for (const [messageId, rawIds] of coverageByMessageId.entries()) {
-          if (rawIds.size !== matchedIds.size) {
-            continue;
-          }
-          const rows = rawBlockSignatureStmt.all(messageId) as Array<{ metadata: string | null }>;
-          if (
-            rows.length === parts.length &&
-            rows.every((row, index) => row.metadata === (parts[index]?.metadata ?? null))
-          ) {
-            alreadyPersisted = true;
-            break;
-          }
-        }
-      }
-
-      const canUseExternalizedFallback = parts.length === 1 || everyPartHasRawBlockId;
-      if (!alreadyPersisted && canUseExternalizedFallback && externalizedTextsById.size > 0) {
-        const externalizedCoverageByMessageId = new Map<number, Set<string>>();
-        for (const rawId of matchedIds) {
-          const externalizedText = externalizedTextsById.get(rawId);
-          if (externalizedText === undefined) {
-            continue;
-          }
-          const externalizedByteSize = Buffer.byteLength(externalizedText, "utf8");
-          const rawIdArgs = [
-            rawId,
-            rawId,
-            rawId,
-            rawId,
-            rawId,
-            rawId,
-            rawId,
-            rawId,
-            rawId,
-            rawId,
-            rawId,
-            rawId,
-            rawId,
-          ];
-          const rows = externalizedCoverageStmt.all(
-            params.conversationId,
-            stored.role,
-            ...rawIdArgs,
-          ) as Array<{
-            messageId: number;
-            fileId: unknown;
-            originalByteSize: unknown;
-            metadata: string | null;
-          }>;
-          for (const row of rows) {
-            if (
-              typeof row.fileId !== "string" ||
-              Number(row.originalByteSize) !== externalizedByteSize
-            ) {
-              continue;
-            }
-            const largeFile = await this.summaryStore.getLargeFile(row.fileId);
-            if (!largeFile) {
-              continue;
-            }
-            let storedText: string;
-            try {
-              storedText = readFileSync(largeFile.storageUri, "utf8");
-            } catch {
-              continue;
-            }
-            if (
-              storedText !== externalizedText ||
-              !externalizedReplayMetadataMatches(row.metadata, parts[0]?.metadata)
-            ) {
-              continue;
-            }
-            const matchedRawIds =
-              externalizedCoverageByMessageId.get(row.messageId) ?? new Set<string>();
-            matchedRawIds.add(rawId);
-            externalizedCoverageByMessageId.set(row.messageId, matchedRawIds);
-          }
-        }
-        alreadyPersisted = Array.from(externalizedCoverageByMessageId.values()).some(
-          (rawIds) => rawIds.size === matchedIds.size,
-        );
-      }
-
-      if (alreadyPersisted) {
-        replayedMessages += 1;
-      } else {
-        filtered.push(message);
-      }
-    }
-
-    if (replayedMessages > 0) {
-      const sessionContext = this.formatSessionLogContext({
-        conversationId: params.conversationId,
-        sessionId: params.sessionId,
-        sessionKey: params.sessionKey,
-      });
-      this.deps.log.warn(
-        `[lcm] ingestBatch: dropped ${replayedMessages}/${params.messages.length} raw-id replay messages for ${sessionContext}`,
-      );
-    }
-
-    return filtered;
-  }
-
-  /**
-   * Existing-conversation bootstrap is a rehydrate path. It may repair small
-   * crash gaps, but it must not replay already persisted transcript rows as
-   * fresh LCM seqs after a runtime re-instantiation.
-   */
-  private async filterBootstrapReplayMessages(params: {
-    messages: AgentMessage[];
-    sessionContext: string;
-    source: string;
-    priorMessages?: AgentMessage[];
-    sessionFile?: string;
-  }): Promise<{ messages: AgentMessage[]; replayGuardExemptPrefixLength: number }> {
-    if (params.messages.length < 3) {
-      return { messages: params.messages, replayGuardExemptPrefixLength: 0 };
-    }
-
-    let replayCandidateLength = 0;
-    while (
-      replayCandidateLength < params.messages.length &&
-      isBootstrapReplayCandidateMessage(params.messages[replayCandidateLength]!)
-    ) {
-      replayCandidateLength += 1;
-    }
-    if (replayCandidateLength < 3) {
-      return { messages: params.messages, replayGuardExemptPrefixLength: 0 };
-    }
-
-    const priorMessages =
-      params.priorMessages ??
-      (params.sessionFile ? await readLeafPathMessages(params.sessionFile) : undefined);
-    if (!priorMessages || priorMessages.length === 0) {
-      return { messages: params.messages, replayGuardExemptPrefixLength: 0 };
-    }
-
-    const replayCandidates = params.messages.slice(0, replayCandidateLength);
-    const earlierReplayCandidates = (
-      params.priorMessages ? priorMessages : priorMessages.slice(0, Math.max(0, priorMessages.length - params.messages.length))
-    ).filter(isBootstrapReplayCandidateMessage);
-    if (earlierReplayCandidates.length < 3) {
-      return { messages: params.messages, replayGuardExemptPrefixLength: 0 };
-    }
-
-    const incomingSignatures = replayCandidates.map(createLosslessMessageSignature);
-    const earlierSignatures = earlierReplayCandidates.map(createLosslessMessageSignature);
-
-    let replayPrefixLength = 0;
-    prefixLoop:
-    for (
-      let candidatePrefixLength = incomingSignatures.length;
-      candidatePrefixLength >= 3;
-      candidatePrefixLength -= 1
-    ) {
-      for (
-        let startIndex = 0;
-        startIndex <= earlierSignatures.length - candidatePrefixLength;
-        startIndex += 1
-      ) {
-        let matched = true;
-        for (let offset = 0; offset < candidatePrefixLength; offset += 1) {
-          if (earlierSignatures[startIndex + offset] !== incomingSignatures[offset]) {
-            matched = false;
-            break;
-          }
-        }
-        if (matched) {
-          replayPrefixLength = candidatePrefixLength;
-          break prefixLoop;
-        }
-      }
-    }
-
-    if (replayPrefixLength > 0) {
-      this.deps.log.warn(
-        `[lcm] bootstrap replay guard: ${params.source} dropped ${replayPrefixLength}/${params.messages.length} replayed transcript messages for ${params.sessionContext}`,
-      );
-    }
-
-    if (replayPrefixLength > 0) {
-      return {
-        messages: params.messages.slice(replayPrefixLength),
-        replayGuardExemptPrefixLength: Math.max(0, replayCandidateLength - replayPrefixLength),
-      };
-    }
-
-    return {
-      messages: params.messages,
-      replayGuardExemptPrefixLength: replayCandidateLength,
-    };
-  }
-
-  private async reconcileTranscriptTailForAfterTurnInSessionQueue(params: {
-    sessionId: string;
-    sessionKey?: string;
-    sessionFile: string;
-    isHeartbeat?: boolean;
-    allowNoAnchorImportOnCheckpointMissing?: boolean;
-  }): Promise<TranscriptReconcileResult> {
-    const queueKey = this.resolveSessionQueueKey(params.sessionId, params.sessionKey);
-    await this.conversationStore.withTransaction(async () => {
-      await this.sessionRolloverDetector.rotateIsolatedCronConversationIfRuntimeChanged({
-        phase: "afterTurn",
-        sessionId: params.sessionId,
-        sessionKey: params.sessionKey,
-        createReplacement: false,
-      });
-      await this.sessionRolloverDetector.rotateStaleSessionKeyConversationIfTrackedTranscriptMissing({
-        phase: "afterTurn",
-        sessionId: params.sessionId,
-        sessionKey: params.sessionKey,
-        sessionFile: params.sessionFile,
-        createReplacement: false,
-      });
-    });
-        const conversation = await this.conversationStore.getConversationForSession({
-          sessionId: params.sessionId,
-          sessionKey: params.sessionKey,
-        });
-        if (!conversation) {
-          if (params.isHeartbeat) {
-            return { importedMessages: 0, blockedByImportCap: false, hasOverlap: true };
-          }
-          // No persisted conversation exists yet. Prefer the transcript over
-          // the runtime delta so foreground prompts that are omitted from
-          // afterTurn's messages array are not lost.
-          let sessionFileState: { size: number } | undefined;
-          try {
-            const sessionFileStats = await stat(params.sessionFile);
-            sessionFileState = { size: sessionFileStats.size };
-          } catch {
-            // Missing files are common for brand-new live sessions; allow the
-            // runtime batch to seed the conversation in that case.
-          }
-          const historicalMessages = await readLeafPathMessages(params.sessionFile);
-          if (historicalMessages.length === 0) {
-            if ((sessionFileState?.size ?? 0) > 0) {
-              this.deps.log.warn(
-                `[lcm] afterTurn: initial transcript read returned no messages from non-empty file; skipping live afterTurn persistence to avoid anchoring past unreadable history session=${params.sessionId}${params.sessionKey?.trim() ? ` sessionKey=${params.sessionKey.trim()}` : ""} sessionFile=${params.sessionFile}`,
-              );
-              return { importedMessages: 0, blockedByImportCap: false, hasOverlap: false };
-            }
-            return { importedMessages: 0, blockedByImportCap: false, hasOverlap: true };
-          }
-          if (batchLooksLikeHeartbeatAckTurn(historicalMessages)) {
-            // Not covered: the runtime batch path owns conversation creation
-            // and heartbeat-ack pruning for brand-new sessions.
-            return { importedMessages: 0, blockedByImportCap: false, hasOverlap: true };
-          }
-          const bootstrapMessages = trimBootstrapMessagesToBudget(
-            historicalMessages,
-            resolveBootstrapMaxTokens(this.config),
-          );
-          if (bootstrapMessages.length === 0) {
-            this.deps.log.warn(
-              `[lcm] afterTurn: initial transcript import exceeded bootstrap budget; skipping live afterTurn persistence to avoid anchoring past unreconciled history session=${params.sessionId}${params.sessionKey?.trim() ? ` sessionKey=${params.sessionKey.trim()}` : ""} sessionFile=${params.sessionFile} sourceMessages=${historicalMessages.length}`,
-            );
-            return { importedMessages: 0, blockedByImportCap: true, hasOverlap: false };
-          }
-          let importedMessages = 0;
-          for (const message of bootstrapMessages) {
-            const result = await this.ingestSingle({
-              sessionId: params.sessionId,
-              sessionKey: params.sessionKey,
-              message,
-              skipReplayTimestampFloodGuard: true,
-            });
-            if (result.ingested) {
-              importedMessages += 1;
-            }
-          }
-          if (importedMessages > 0) {
-            const activeConversation = await this.conversationStore.getConversationForSession({
-              sessionId: params.sessionId,
-              sessionKey: params.sessionKey,
-            });
-            if (activeConversation) {
-              this.recordRecentBootstrapImport(
-                activeConversation.conversationId,
-                importedMessages,
-                "imported initial afterTurn transcript",
-              );
-              await this.refreshBootstrapState({
-                conversationId: activeConversation.conversationId,
-                sessionFile: params.sessionFile,
-              });
-            }
-          }
-          return {
-            importedMessages,
-            blockedByImportCap: bootstrapMessages.length < historicalMessages.length,
-            hasOverlap: true,
-            transcriptCovered: true,
-          };
-        }
-
-        // OpenClaw can submit the foreground prompt outside the mutable
-        // messages array passed to afterTurn. The transcript has the complete
-        // turn by this point, so reconcile it before accepting assistant-only
-        // deltas from the runtime snapshot.
-        const checkpoint = await this.summaryStore.getConversationBootstrapState(
-          conversation.conversationId,
-        );
-        let sessionFileState: { size: number; mtimeMs: number } | undefined;
-        let sessionFileStatError: unknown;
-        try {
-          const sessionFileStats = await stat(params.sessionFile);
-          sessionFileState = {
-            size: sessionFileStats.size,
-            mtimeMs: Math.trunc(sessionFileStats.mtimeMs),
-          };
-        } catch (error) {
-          sessionFileStatError = error;
-          // Leave undefined: without stat proof, do not use append-only guards or slow-read caps.
-        }
-        const transcriptEpochShrank = checkpointIsPastTranscriptEof(
-          checkpoint,
-          sessionFileState?.size ?? Number.POSITIVE_INFINITY,
-        );
-        if (
-          checkpoint &&
-          checkpoint.sessionFilePath === params.sessionFile &&
-          checkpoint.lastProcessedOffset >= 0 &&
-          !transcriptEpochShrank
-        ) {
-          const appended = await readAppendedLeafPathMessages({
-            sessionFile: params.sessionFile,
-            offset: checkpoint.lastProcessedOffset,
-          });
-          if (appended.canUseAppendOnly) {
-            const placeholderCheckpoint =
-              checkpoint.lastSeenSize === 0 &&
-              checkpoint.lastSeenMtimeMs === 0 &&
-              checkpoint.lastProcessedOffset === 0 &&
-              checkpoint.lastProcessedEntryHash === null;
-            if (params.isHeartbeat) {
-              if (!placeholderCheckpoint) {
-                await this.refreshBootstrapState({
-                  conversationId: conversation.conversationId,
-                  sessionFile: params.sessionFile,
-                });
-                this.deps.log.debug(
-                  `[lcm] afterTurn: skipped heartbeat transcript append-only delta and refreshed checkpoint conversation=${conversation.conversationId} sessionFile=${params.sessionFile} appendedMessages=${appended.messages.length}`,
-                );
-              }
-              // Heartbeat turns are never persisted; the append-only delta is
-              // intentionally skipped, so the transcript counts as covered.
-              return {
-                importedMessages: 0,
-                blockedByImportCap: false,
-                hasOverlap: true,
-                transcriptCovered: true,
-              };
-            }
-            if (placeholderCheckpoint && appended.messages.length > 0) {
-              const reconcile = await this.reconcileSessionTail({
-                sessionId: params.sessionId,
-                sessionKey: params.sessionKey,
-                conversationId: conversation.conversationId,
-                historicalMessages: appended.messages,
-                noAnchorImportReason: "placeholder-checkpoint-recovery",
-              });
-              if (reconcile.importedMessages > 0) {
-                this.recordRecentBootstrapImport(
-                  conversation.conversationId,
-                  reconcile.importedMessages,
-                  "reconciled missing session messages",
-                );
-                await this.refreshBootstrapState({
-                  conversationId: conversation.conversationId,
-                  sessionFile: params.sessionFile,
-                });
-              }
-              return {
-                ...reconcile,
-                transcriptCovered: reconcile.hasOverlap || reconcile.importedMessages > 0,
-              };
-            }
-
-            const appendOnlySessionContext = this.formatSessionLogContext({
-              conversationId: conversation.conversationId,
-              sessionId: params.sessionId,
-              sessionKey: params.sessionKey,
-            });
-            const replayFiltered = await this.filterBootstrapReplayMessages({
-              messages: appended.messages,
-              sessionContext: appendOnlySessionContext,
-              source: "afterTurn transcript reconcile append-only",
-              sessionFile: params.sessionFile,
-            });
-            const replayFilteredMessages = replayFiltered.messages;
-            const appendOnlyOverlapsPersisted = await this.appendOnlyMessagesOverlapPersistedTranscript({
-              conversationId: conversation.conversationId,
-              messages: replayFilteredMessages,
-              unfilteredMessages: appended.messages,
-              sessionContext: appendOnlySessionContext,
-              source: "afterTurn transcript reconcile append-only",
-            });
-            if (!appendOnlyOverlapsPersisted) {
-              let importedMessages = 0;
-              for (const [index, message] of replayFilteredMessages.entries()) {
-                const result = await this.ingestSingle({
-                  sessionId: params.sessionId,
-                  sessionKey: params.sessionKey,
-                  message,
-                  skipReplayTimestampFloodGuard:
-                    index < replayFiltered.replayGuardExemptPrefixLength,
-                });
-                if (result.ingested) {
-                  importedMessages += 1;
-                }
-              }
-              if (importedMessages > 0) {
-                this.recordRecentBootstrapImport(
-                  conversation.conversationId,
-                  importedMessages,
-                  "reconciled missing session messages",
-                );
-                await this.refreshBootstrapState({
-                  conversationId: conversation.conversationId,
-                  sessionFile: params.sessionFile,
-                });
-              }
-              return {
-                importedMessages,
-                blockedByImportCap: false,
-                hasOverlap: true,
-                transcriptCovered: true,
-              };
-            }
-          }
-        }
-
-        // Slow path: checkpoint missing, path mismatched, or non-append-only.
-        // Cap full re-reads only for unchanged file states. If the transcript
-        // changed since the last full read, reconcile again; if it did not,
-        // skip without advancing the checkpoint so stale state can be retried
-        // after a later file change, process restart, or cap eviction.
-        const fullReadKey = `${queueKey}\u0000${params.sessionFile}`;
-        const reason = !checkpoint
-          ? "checkpoint-missing"
-          : checkpoint.sessionFilePath !== params.sessionFile
-            ? "path-mismatch"
-            : transcriptEpochShrank
-              ? "same-path-shrink"
-              : "append-only-ineligible";
-        if (reason === "same-path-shrink") {
-          this.afterTurnReconcileFullReadStates.delete(fullReadKey);
-        }
-        const rememberedFileState = this.afterTurnReconcileFullReadStates.get(fullReadKey);
-        if (
-          rememberedFileState
-          && sessionFileState
-          && rememberedFileState.size === sessionFileState.size
-          && rememberedFileState.mtimeMs === sessionFileState.mtimeMs
-        ) {
-          this.deps.log.debug(
-            `[lcm] afterTurn: transcript reconcile slow path skipped (file state already read this process) conversation=${conversation.conversationId} reason=${reason} sessionFile=${params.sessionFile}`,
-          );
-          // The memo is only populated after a successful covered full read,
-          // and the file has not changed since — still covered.
-          return {
-            importedMessages: 0,
-            blockedByImportCap: false,
-            hasOverlap: true,
-            transcriptCovered: true,
-          };
-        }
-
-        const rememberSlowReadState = (): void => {
-          if (!sessionFileState) {
-            return;
-          }
-          if (
-            !this.afterTurnReconcileFullReadStates.has(fullReadKey)
-            && this.afterTurnReconcileFullReadStates.size
-              >= LcmContextEngine.AFTER_TURN_RECONCILE_KEY_CAP
-          ) {
-            const oldest = this.afterTurnReconcileFullReadStates.keys().next().value;
-            if (typeof oldest === "string") {
-              this.afterTurnReconcileFullReadStates.delete(oldest);
-            }
-          }
-          this.afterTurnReconcileFullReadStates.set(fullReadKey, sessionFileState);
-        };
-        const slowPathStartedAt = Date.now();
-
-        if (isMissingFileError(sessionFileStatError)) {
-          if (!checkpoint) {
-            try {
-              await this.summaryStore.upsertConversationBootstrapState({
-                conversationId: conversation.conversationId,
-                sessionFilePath: params.sessionFile,
-                lastSeenSize: 0,
-                lastSeenMtimeMs: 0,
-                lastProcessedOffset: 0,
-                lastProcessedEntryHash: null,
-              });
-            } catch (seedError) {
-              this.deps.log.warn(
-                `[lcm] afterTurn: transcript reconcile slow path failed to seed placeholder bootstrap_state conversation=${conversation.conversationId} sessionFile=${params.sessionFile} error=${seedError instanceof Error ? seedError.message : String(seedError)}`,
-              );
-            }
-            this.deps.log.warn(
-              `[lcm] afterTurn: session file missing; skipping transcript reconcile full reread; could not stat/read transcript; allowing live afterTurn persistence and seeding placeholder bootstrap_state at offset=0 to unblock next-turn recovery conversation=${conversation.conversationId} reason=${reason} sessionFile=${params.sessionFile}`,
-            );
-          } else {
-            this.deps.log.warn(
-              `[lcm] afterTurn: session file missing; skipping transcript reconcile full reread; preserving existing checkpoint (offset=${checkpoint.lastProcessedOffset}) conversation=${conversation.conversationId} reason=${reason} sessionFile=${params.sessionFile}`,
-            );
-          }
-          return {
-            importedMessages: 0,
-            blockedByImportCap: false,
-            hasOverlap: true,
-          };
-        }
-
-        // Distinguish empty-file from read/parse error: stat the file and
-        // only treat it as "actually empty" when size is 0. A non-zero file
-        // returning empty `historicalMessages` indicates the parser hit an
-        // error (and `readLeafPathMessages` swallows those into `[]`); in
-        // that case we must NOT mark the bootstrap checkpoint as fully
-        // processed, otherwise future afterTurns will skip reconciliation
-        // and we lose messages.
-        const historicalMessages = await readLeafPathMessages(params.sessionFile);
-        if (reason === "path-mismatch") {
-          const ambiguousRollover =
-            await this.sessionRolloverDetector.findAmbiguousSessionKeyRuntimeRollover({
-              phase: "afterTurn",
-              sessionId: params.sessionId,
-              sessionKey: params.sessionKey,
-              sessionFile: params.sessionFile,
-            });
-          if (ambiguousRollover) {
-            const activeBootstrapState =
-              await this.summaryStore.getConversationBootstrapState(
-                ambiguousRollover.conversationId,
-              );
-            const hasFrontierAnchor =
-              await this.sessionRolloverDetector.transcriptContainsCurrentConversationTailAnchor({
-                conversationId: ambiguousRollover.conversationId,
-                historicalMessages,
-                checkpointEntryHash: activeBootstrapState?.lastProcessedEntryHash,
-              });
-            if (!hasFrontierAnchor) {
-              // Tier-2 resolution: archive the frozen conversation and
-              // create the replacement now. This turn's persistence is
-              // skipped (unsafe-to-advance), and the next turn reconciles
-              // the full transcript into the fresh conversation.
-              const rotatedForFreshTranscript =
-                await this.sessionRolloverDetector.rotateAmbiguousRolloverForProvablyFreshTranscript({
-                  phase: "afterTurn",
-                  sessionId: params.sessionId,
-                  rollover: ambiguousRollover,
-                  candidateMessages: historicalMessages,
-                  createReplacement: true,
-                });
-              if (rotatedForFreshTranscript) {
-                return {
-                  importedMessages: 0,
-                  blockedByImportCap: false,
-                  blockedReason: "ambiguous-rollover-rotated-fresh-transcript",
-                  hasOverlap: false,
-                };
-              }
-              this.sessionRolloverDetector.logAmbiguousSessionKeyRuntimeRollover({
-                phase: "afterTurn",
-                rollover: ambiguousRollover,
-                sessionId: params.sessionId,
-                sessionFile: params.sessionFile,
-              });
-              return {
-                importedMessages: 0,
-                blockedByImportCap: false,
-                blockedReason: "ambiguous-session-key-runtime-rollover",
-                hasOverlap: false,
-              };
-            }
-          }
-        }
-        if (historicalMessages.length === 0) {
-          if (!sessionFileState) {
-            // #649 added this permissive stat-fail fallback expecting the
-            // afterTurn-tail `refreshAfterTurnBootstrapState` hook to refresh
-            // the checkpoint. That hook delegates to refreshBootstrapState,
-            // which itself calls `stat(sessionFile)` and throws on failure;
-            // the hook's catch then logs a warn and leaves
-            // conversation_bootstrap_state NULL. Subsequent turns re-enter
-            // the slow path with reason="checkpoint-missing" (excluded from
-            // allowNoAnchorImport) and the conversation gets stuck in a
-            // transparent-passthrough state where compaction never runs.
-            //
-            // Seed a placeholder bootstrap_state row ONLY when no checkpoint
-            // already exists. If a valid checkpoint is present (with a
-            // non-zero offset), a transient stat/read failure must NOT reset
-            // it to zero — that would cause the next successful read to
-            // replay every message from offset=0, duplicating rows in the
-            // messages table (identity_hash is not a uniqueness guard).
-            if (!checkpoint) {
-              try {
-                await this.summaryStore.upsertConversationBootstrapState({
-                  conversationId: conversation.conversationId,
-                  sessionFilePath: params.sessionFile,
-                  lastSeenSize: 0,
-                  lastSeenMtimeMs: 0,
-                  lastProcessedOffset: 0,
-                  lastProcessedEntryHash: null,
-                });
-              } catch (seedError) {
-                this.deps.log.warn(
-                  `[lcm] afterTurn: transcript reconcile slow path failed to seed placeholder bootstrap_state conversation=${conversation.conversationId} sessionFile=${params.sessionFile} error=${seedError instanceof Error ? seedError.message : String(seedError)}`,
-                );
-              }
-              this.deps.log.warn(
-                `[lcm] afterTurn: transcript reconcile slow path could not stat/read transcript; allowing live afterTurn persistence and seeding placeholder bootstrap_state at offset=0 to unblock next-turn recovery conversation=${conversation.conversationId} sessionFile=${params.sessionFile}`,
-              );
-            } else {
-              // Checkpoint exists with a valid offset — a transient stat/read
-              // failure must NOT overwrite it. Leave the existing checkpoint
-              // intact so the next successful read resumes from the right offset.
-              this.deps.log.warn(
-                `[lcm] afterTurn: transcript reconcile slow path could not stat/read transcript; preserving existing checkpoint (offset=${checkpoint.lastProcessedOffset}) instead of reseeding conversation=${conversation.conversationId} sessionFile=${params.sessionFile}`,
-              );
-            }
-            return {
-              importedMessages: 0,
-              blockedByImportCap: false,
-              hasOverlap: true,
-            };
-          }
-          if (sessionFileState.size === 0) {
-            // File is genuinely empty — refresh the checkpoint so the next
-            // afterTurn takes the incremental path.
-            await this.refreshBootstrapState({
-              conversationId: conversation.conversationId,
-              sessionFile: params.sessionFile,
-            });
-            rememberSlowReadState();
-          } else {
-            this.deps.log.warn(
-              `[lcm] afterTurn: transcript reconcile slow path read empty messages from non-empty file (${sessionFileState?.size ?? "?"} bytes) — skipping checkpoint refresh to avoid dropping messages on parser failure conversation=${conversation.conversationId} sessionFile=${params.sessionFile}`,
-            );
-          }
-          // An empty transcript cannot carry the turn — let the runtime
-          // batch persist it (transcriptCovered stays false).
-          return {
-            importedMessages: 0,
-            blockedByImportCap: false,
-            hasOverlap: sessionFileState.size === 0,
-          };
-        }
-        // #837: a conversation with bootstrapped_at SET but no bootstrap_state
-        // row reaches reason="checkpoint-missing" with a non-anchoring frontier
-        // (e.g. a single injected metadata preamble). Without a no-anchor import
-        // it imports 0 messages and never persists a checkpoint, so afterTurn
-        // loops the "did not cover the transcript frontier" warning forever and
-        // compaction never runs. The rotate lane already recovers via
-        // allowNoAnchorImportOnCheckpointMissing; mirror that on the afterTurn
-        // lane, but ONLY for the observed injected-metadata frontier. A real
-        // historical DB tail with a divergent rewritten transcript must still
-        // freeze per #649's no-proof-no-advance guard, so do not treat
-        // bootstrapped_at alone as lineage proof. The downstream no-anchor
-        // import path is itself guarded (replay-overlap detection, import cap,
-        // delivery-only block).
-        let checkpointMissingMetadataFrontier = false;
-        if (
-          reason === "checkpoint-missing" &&
-          conversation.sessionId === params.sessionId &&
-          conversation.bootstrappedAt !== null
-        ) {
-          const [existingMessageCount, latestPersistedMessage] = await Promise.all([
-            this.conversationStore.getMessageCount(conversation.conversationId),
-            this.conversationStore.getLastMessage(conversation.conversationId),
-          ]);
-          checkpointMissingMetadataFrontier =
-            existingMessageCount === 1 &&
-            latestPersistedMessage !== null &&
-            isLikelyInjectedMetadataPreambleRecord(latestPersistedMessage);
-        }
-        const recoverCheckpointMissingNoAnchor =
-          reason === "checkpoint-missing" &&
-          (params.allowNoAnchorImportOnCheckpointMissing === true ||
-            checkpointMissingMetadataFrontier);
-        // A transcript whose session header id differs from the checkpoint's
-        // is a *declared* epoch change (rewrite/rotation) — no heuristics
-        // needed, so a same-path full rewrite may import as a new epoch
-        // instead of freezing on "no anchor". The no-anchor path's replay
-        // guards and import caps still apply as sanity bounds.
-        const transcriptHeader = await readTranscriptHeader(params.sessionFile);
-        const declaredEpochRollover =
-          resolveEpochRoute({
-            checkpointHeaderId: checkpoint?.sessionHeaderId,
-            transcriptHeaderId: transcriptHeader.sessionHeaderId,
-          }) === "declared-rollover";
-        if (declaredEpochRollover) {
-          this.deps.log.warn(
-            `[lcm] afterTurn: transcript session header changed (${checkpoint?.sessionHeaderId} -> ${transcriptHeader.sessionHeaderId}); treating as declared epoch rollover conversation=${conversation.conversationId} reason=${reason} sessionFile=${params.sessionFile}`,
-          );
-        }
-        const reconcile = await this.reconcileSessionTail({
-          sessionId: params.sessionId,
-          sessionKey: params.sessionKey,
-          conversationId: conversation.conversationId,
-          historicalMessages,
-          lastProcessedEntryId: declaredEpochRollover
-            ? null
-            : checkpoint?.lastProcessedEntryId ?? null,
-          skipContentAnchorScan: reason === "same-path-shrink",
-          allowNoAnchorImport:
-            reason === "path-mismatch" ||
-            reason === "same-path-shrink" ||
-            declaredEpochRollover ||
-            recoverCheckpointMissingNoAnchor,
-          noAnchorImportReason: recoverCheckpointMissingNoAnchor
-            ? params.allowNoAnchorImportOnCheckpointMissing === true
-              ? "rotate-checkpoint-missing"
-              : "checkpoint-missing-recovery"
-            : declaredEpochRollover && reason === "append-only-ineligible"
-              ? "declared-epoch-rollover"
-              : reason,
-        });
-        if (reconcile.blockedByImportCap) {
-          // Capped passes import a bounded chunk of anchored backlog; report
-          // the partial progress while leaving the checkpoint un-advanced so
-          // the next pass continues the drain.
-          return {
-            importedMessages: reconcile.importedMessages,
-            blockedByImportCap: true,
-            hasOverlap: reconcile.hasOverlap,
-          };
-        }
-        if (reconcile.importedMessages > 0) {
-          this.recordRecentBootstrapImport(
-            conversation.conversationId,
-            reconcile.importedMessages,
-            "reconciled missing session messages",
-          );
-        }
-        if (!reconcile.hasOverlap && reconcile.importedMessages === 0) {
-          this.deps.log.warn(
-            `[lcm] afterTurn: transcript reconcile found no anchor and imported 0 messages; skipping checkpoint refresh conversation=${conversation.conversationId} reason=${reason} sessionFile=${params.sessionFile} historicalMessages=${historicalMessages.length}`,
-          );
-          return { importedMessages: 0, blockedByImportCap: false, hasOverlap: false };
-        }
-        // Refresh only after the slow-path read either found an overlap or
-        // imported the bounded no-anchor epoch. A no-overlap/no-import result
-        // leaves the checkpoint stale on purpose so future turns can retry.
-        await this.refreshBootstrapState({
-          conversationId: conversation.conversationId,
-          sessionFile: params.sessionFile,
-        });
-        rememberSlowReadState();
-        this.deps.log.warn(
-          `[lcm] afterTurn: transcript reconcile slow path (full re-read) conversation=${conversation.conversationId} reason=${reason} sessionFile=${params.sessionFile} historicalMessages=${historicalMessages.length} importedMessages=${reconcile.importedMessages} duration=${formatDurationMs(Date.now() - slowPathStartedAt)}`,
-        );
-        return {
-          importedMessages: reconcile.importedMessages,
-          blockedByImportCap: false,
-          hasOverlap: reconcile.hasOverlap,
-          transcriptCovered: true,
-        };
-  }
-
-  private filterSyntheticHeartbeatTranscriptMessages(params: {
-    messages: AgentMessage[];
-    sessionContext: string;
-    source: string;
-  }): AgentMessage[] {
-    const filtered = filterSyntheticHeartbeatMessages(params.messages);
-    if (filtered.skipped > 0) {
-      this.deps.log.debug(
-        `[lcm] ${params.source}: skipped ${filtered.skipped}/${params.messages.length} synthetic heartbeat transcript messages for ${params.sessionContext}`,
-      );
-    }
-    return filtered.messages;
-  }
-
-  private async reconcileTranscriptTailForAfterTurn(params: {
-    sessionId: string;
-    sessionKey?: string;
-    sessionFile: string;
-    isHeartbeat?: boolean;
-    allowNoAnchorImportOnCheckpointMissing?: boolean;
-  }): Promise<TranscriptReconcileResult> {
-    const queueKey = this.resolveSessionQueueKey(params.sessionId, params.sessionKey);
-    return await this.withSessionQueue(
-      queueKey,
-      () => this.reconcileTranscriptTailForAfterTurnInSessionQueue(params),
-      {
-        operationName: "afterTurnTranscriptReconcile",
-        context: [
-          `session=${params.sessionId}`,
-          ...(params.sessionKey?.trim() ? [`sessionKey=${params.sessionKey.trim()}`] : []),
-        ].join(" "),
-      },
-    );
-  }
-
-  /**
-   * Persist bootstrap checkpoint metadata anchored to the current DB frontier.
-   *
-   * By default, the frontier hash follows the latest persisted DB message. The
-   * first-time bootstrap path can override it with the raw transcript hash so
-   * later reconciliation can anchor entries whose DB content was externalized.
-   */
-  private async refreshBootstrapState(params: {
-    conversationId: number;
-    sessionFile: string;
-    fileStats?: { size: number; mtimeMs: number };
-    lastProcessedEntryHash?: string | null;
-    forkBounded?: boolean;
-    forkSourceMessageCount?: number;
-  }): Promise<void> {
-    const latestDbMessage = await this.conversationStore.getLastMessage(params.conversationId);
-    const fileStats = params.fileStats ?? (await stat(params.sessionFile));
-    // The checkpoint marks the whole file processed (offset = size), so the
-    // exact resume anchor is the envelope id of the file's last message entry.
-    let lastProcessedEntryId: string | null = null;
-    const lastEntryLine = await readLastJsonlEntryBeforeOffset(
-      params.sessionFile,
-      fileStats.size,
-      true,
-    );
-    if (lastEntryLine) {
-      try {
-        const parsed = JSON.parse(lastEntryLine) as { id?: unknown; uuid?: unknown };
-        const rawId = typeof parsed.id === "string" ? parsed.id : typeof parsed.uuid === "string" ? parsed.uuid : "";
-        lastProcessedEntryId = rawId.trim() || null;
-      } catch {
-        // Bare-message lines have no envelope id.
-      }
-    }
-    const header = await readTranscriptHeader(params.sessionFile);
-    await this.summaryStore.upsertConversationBootstrapState({
-      conversationId: params.conversationId,
-      sessionFilePath: params.sessionFile,
-      lastSeenSize: fileStats.size,
-      lastSeenMtimeMs: Math.trunc(fileStats.mtimeMs),
-      lastProcessedOffset: fileStats.size,
-      lastProcessedEntryHash:
-        params.lastProcessedEntryHash !== undefined
-          ? params.lastProcessedEntryHash
-          : latestDbMessage
-            ? createBootstrapEntryHash({
-                role: latestDbMessage.role,
-                content: latestDbMessage.content,
-                tokenCount: latestDbMessage.tokenCount,
-              })
-            : null,
-      sessionHeaderId: header.sessionHeaderId,
-      lastProcessedEntryId,
-      forkBounded: params.forkBounded,
-      forkSourceMessageCount: params.forkSourceMessageCount,
-    });
-  }
 
 
   async bootstrap(params: {
@@ -3589,7 +1664,7 @@ export class LcmContextEngine implements ContextEngine {
               forkSourceMessageCount: number;
             },
           ): Promise<void> => {
-            await this.refreshBootstrapState({
+            await this.transcriptReconciler.refreshBootstrapState({
               conversationId,
               sessionFile: params.sessionFile,
               fileStats: {
@@ -3797,18 +1872,18 @@ export class LcmContextEngine implements ContextEngine {
                   sessionId: params.sessionId,
                   sessionKey: params.sessionKey,
                 });
-                const replayFiltered = await this.filterBootstrapReplayMessages({
+                const replayFiltered = await this.transcriptReconciler.filterBootstrapReplayMessages({
                   messages: appended.messages,
                   sessionContext: appendOnlySessionContext,
                   source: "bootstrap append-only",
                   sessionFile: params.sessionFile,
                 });
-                const replayFilteredMessages = this.filterSyntheticHeartbeatTranscriptMessages({
+                const replayFilteredMessages = this.transcriptReconciler.filterSyntheticHeartbeatTranscriptMessages({
                   messages: replayFiltered.messages,
                   sessionContext: appendOnlySessionContext,
                   source: "bootstrap append-only",
                 });
-                const appendOnlyOverlapsPersisted = await this.appendOnlyMessagesOverlapPersistedTranscript({
+                const appendOnlyOverlapsPersisted = await this.transcriptReconciler.appendOnlyMessagesOverlapPersistedTranscript({
                   conversationId,
                   messages: replayFilteredMessages,
                   unfilteredMessages: appended.messages,
@@ -3955,7 +2030,7 @@ export class LcmContextEngine implements ContextEngine {
 
           // Existing conversation path: reconcile crash gaps by appending JSONL
           // messages that were never persisted to LCM.
-          const reconcile = await this.reconcileSessionTail({
+          const reconcile = await this.transcriptReconciler.reconcileSessionTail({
             sessionId: params.sessionId,
             sessionKey: params.sessionKey,
             conversationId,
@@ -4039,7 +2114,7 @@ export class LcmContextEngine implements ContextEngine {
         if (conversation) {
           const pruned = await pruneHeartbeatOkTurns(this.conversationStore, conversation.conversationId);
           if (pruned > 0) {
-            await this.refreshBootstrapState({
+            await this.transcriptReconciler.refreshBootstrapState({
               conversationId: conversation.conversationId,
               sessionFile: params.sessionFile,
             });
@@ -4301,7 +2376,7 @@ export class LcmContextEngine implements ContextEngine {
 
         if (result.changed) {
           try {
-            await this.refreshBootstrapState({
+            await this.transcriptReconciler.refreshBootstrapState({
               conversationId: conversation.conversationId,
               sessionFile: params.sessionFile,
             });
@@ -4556,7 +2631,7 @@ export class LcmContextEngine implements ContextEngine {
               sessionKey: params.sessionKey,
             });
             if (conversation) {
-              messages = await this.filterPersistedRawIdReplayBatch({
+              messages = await this.transcriptReconciler.filterPersistedRawIdReplayBatch({
                 conversationId: conversation.conversationId,
                 sessionId: params.sessionId,
                 sessionKey: params.sessionKey,
@@ -4643,7 +2718,7 @@ export class LcmContextEngine implements ContextEngine {
       hasOverlap: true,
     };
     try {
-      transcriptReconcileResult = await this.reconcileTranscriptTailForAfterTurn({
+      transcriptReconcileResult = await this.transcriptReconciler.reconcileTranscriptTailForAfterTurn({
         sessionId: params.sessionId,
         sessionKey: params.sessionKey,
         sessionFile: params.sessionFile,
@@ -4787,7 +2862,7 @@ export class LcmContextEngine implements ContextEngine {
               sessionKey: params.sessionKey,
             });
             try {
-              await this.refreshBootstrapState({
+              await this.transcriptReconciler.refreshBootstrapState({
                 conversationId: conversation.conversationId,
                 sessionFile: params.sessionFile,
               });
@@ -4854,7 +2929,7 @@ export class LcmContextEngine implements ContextEngine {
     }
     const refreshAfterTurnBootstrapState = async (): Promise<void> => {
       try {
-        await this.refreshBootstrapState({
+        await this.transcriptReconciler.refreshBootstrapState({
           conversationId: conversation.conversationId,
           sessionFile: params.sessionFile,
         });
